Inital commit
This contains the examples from the presentation about ES2016 classes that was
held on the sthlm.js meetup on the 2016-03-01.

There are three classes, `Bazar`, `Person` and `Fool` that extends from
`Person`. These are written in four different styles, using more and more
ES2015 features, up til the `class` syntax.